5/2_12:44

+++ OCR/fb_project/fbclub_lib.py
@@ -1,15 +1,16 @@
-from cv2 import cv2
 import numpy as np
 import pyautogui
 import win32gui
 import win32con
 import win32com.client
+import sys,os,re,json,time
+import pytesseract
+import jieba,joblib
+from cv2 import cv2
 from PyQt5.QtWidgets import QApplication
 from PyQt5.QtGui import *
 from PIL import ImageGrab,Image,ImageEnhance
-import sys,os,re,json,time
-import pytesseract
-import jieba
+from sklearn.preprocessing import  PolynomialFeatures
 
 def click_image(image,pos,  action, timestamp,offset=5):
     img = cv2.imread(image)
@@ -28,7 +29,7 @@ def imagesearch(image, precision=0.8):
     res = cv2.matchTemplate(img_gray, template, cv2.TM_CCOEFF_NORMED)
     min_val, max_val, min_loc, max_loc = cv2.minMaxLoc(res)
     if max_val < precision:
-        return [800,500]
+        return [300,500]
     return max_loc #返回圖片座標
 
 def is_number(s):
@@ -44,6 +45,20 @@ def is_number(s):
     except (TypeError, ValueError):
         pass
     return False
+    
+def check_valid_nlp_value_num(usr_dict,target_nums):
+    count=0
+    for key ,value in usr_dict.items():
+        if value!='null':
+            count+=1
+    if count < target_nums:
+        usr_dict.clear()
+
+def check_valid_dict_value_num(usr_dict,target_nums):
+    if len(usr_dict) < target_nums:
+        usr_dict.clear()
+    if '' in usr_dict.values():
+        usr_dict.clear()
 
 class project :
     def __init__(self,hwnd,filename,left,up,right,down,path):
@@ -55,65 +70,49 @@ class project :
         self.down=down
         newpath=str(path)+"/"
         self.path=newpath
+        self.format_number = 0      #房間總數
         jieba.initialize()
         jieba.load_userdict("./jieba/district.txt")
-    def check_valid_nlp_value_num(usr_dict,target_nums):
-        count=0
-        for key ,value in usr_dict.items():
-            if value!='null':
-                count+=1
-        if count < target_nums:
-            usr_dict.clear()
-    def check_valid_dict_value_num(usr_dict,target_nums):
-        if len(usr_dict) < target_nums:
-            usr_dict.clear()
-    def write_json(self,data,filename='./text_file/text.json'):
-        data=json.load(open(filename))
-        if type(data) is dict:
-            data = [data]
-        data.append(a)
-        with open(filename, 'w') as outfile:
-            json.dump(data, outfile,indent=4)
+
     def web_screenshot(self):            #web截圖 圖名稱screenshot
         shell = win32com.client.Dispatch("WScript.Shell")
         shell.SendKeys('%')
         win32gui.SetForegroundWindow(self.hwnd)
         img = ImageGrab.grab()
-        img.save("./screenshot/temporary.jpg")
+        img.save("temporary.jpg")
 
     def divid(self):      #切割成特定範圍的圖片
-        img = Image.open("./screenshot/temporary.jpg")
-        new_mg = img.crop((self.left,self.up,self.right,self.down))
-        enh_con = ImageEnhance.Contrast(new_mg)
+        img = Image.open("temporary.jpg")
+        self.new_mg = img.crop((self.left,self.up,self.right,self.down))
+        enh_con = ImageEnhance.Contrast(self.new_mg)
         contrast=2
         image_contrasted = enh_con.enhance(contrast)
-        image_contrasted.save("./screenshot/"+self.filename+".jpg")
-        os.remove("./screenshot/temporary.jpg")
+        image_contrasted.save(self.filename+".jpg")
+        os.remove("temporary.jpg")
 
     def ocr(self):                           #ocr影像辨識
-        test_img = Image.open("./screenshot/"+self.filename+".jpg")
+        test_img = Image.open(self.filename+".jpg")
         newtext = pytesseract.image_to_string(test_img, lang='chi_tra+fbclub')
         fixed_text = newtext.strip()
         self.text=fixed_text
-        print(self.text)
 
     def cut_word(self):  #斷詞取得目標
         temporary = './text_file/temporary.txt'
         keywords = []
         self.localdic = dict()
         form={
-                    "price":["價格","開價","售價","總價"],
-                    "location":["地址","地點","區域"],
-                    "size":["建坪","坪數","總建"],
-                    "age":["屋齡"],
-                    "format":["格局"],	
+                "price":["價格","開價","售價","總價"],
+                "location":["地址","地點","區域"],
+                "size":["建坪","坪數","總建"],
+                "age":["屋齡"],
+                "format":["格局"],	
         }
         regularform={
-                    "price":"(\d+)萬",
-                    "location":"(\D+?)[區]",
-                    "size":"(\d*[\.\d]*)[坪]*",
-                    "age":"(\d+)年",
-                    "format":"(\d+)[房/]*(\d+)[廳/]*(\d+)[衛]*",
+                "price":"(\d+)萬",
+                "location":"(\D+?)[區]",
+                "size":"(\d*[\.\d]*)[坪]*",
+                "age":"(\d+)年",
+                "format":"(\d+)[房/]*(\d+)[廳/]*(\d+)[衛]*",
         }
         with open(temporary,'w',encoding='utf-8') as t:       #temporary檔寫入ocr檔案
             t.writelines(self.text.replace(" ",""))
@@ -125,7 +124,7 @@ class project :
                 if ':' in sentence:
                     session = sentence.strip().split(":")[0]
                     contain = sentence.strip().split(":")[1]
-                    cut_session = jieba.cut(session)
+                    cut_session = jieba.lcut(session)
                     for word in cut_session:
                         if word in keywords:
                             for name,data in form.items():      #正規化
@@ -136,7 +135,6 @@ class project :
                                                 if len(re.compile(regularform[name]).findall(contain)):
                                                     localformat=re.compile(regularform[name]).findall(contain)[0]
                                                     a=''
-                                                    self.format_number = 0      #房間總數
                                                     for rom in localformat:
                                                         a += rom+'/'
                                                         self.format_number += int(rom)
@@ -146,7 +144,7 @@ class project :
                                             except Exception as e:
                                                 print("exception :")
                                                 print(e)
-                                        if name=='location':
+                                        elif name=='location':
                                             try:
                                                 loc=jieba.lcut(contain)
                                                 for text in loc:
@@ -154,15 +152,14 @@ class project :
                                                         self.localdic[name]=re.compile(regularform['location']).findall(text)[0]+"區"
                                             except Exception as e:
                                                 print("exception :")
-                                                print(e)																			
+                                                print(e)																
                                         else:
                                             if len(re.compile(regularform[name]).findall(contain)):
                                                 self.localdic[name]=re.compile(regularform[name]).findall(contain)[0]
                                             else:
                                                 self.localdic[name]=''.join(re.compile(regularform[name]).findall(contain))
-        check_valid_dict_value_num(self.localdic,4)
-        print(self.localdic["location"])
-        print(type(self.localdic["location"]))
+
+        check_valid_dict_value_num(self.localdic,5)
 
     def jsons_nlp(self):
         temporary='./text_file/temporary.txt'
@@ -210,7 +207,6 @@ class project :
                 input_ids = inputs["input_ids"].tolist()[0]
 
                 context_tokens = tokenizer.convert_ids_to_tokens(input_ids)
-
                 logits = model(**inputs,return_dict=True)
                 answer_start_scores = logits['start_logits'] 
                 answer_end_scores = logits['end_logits']
@@ -229,29 +225,60 @@ class project :
                 print(f"Question: {question}")
                 print(f"Answer not found")
         print("--- %s seconds ---" % (time.time() - start_time))
-        check_valid_nlp_value_num(self.nlpdic,target_nums)
+        check_valid_nlp_value_num(self.nlpdic,5)
         print (self.nlpdic)
 
-    def judge(self,index,index1,index2): 
-        temporary='./text_file/temporary.txt'
 
-        #取出dictionary的值放入預測模型，判斷是否符合標準，有則保留dict
-        #放入index,index1,index2
-        #if放入index1，index1.append(self.localdic)
-        #else，index2.append(self.localdic)
-        #index1,index2宣告在fbclub_project的start函式
-        index.append(self.localdic)
-        index1.append(self.localdic)
-        index2.append(self.localdic)
+    def judge(self,index,index1,index2,num1,num2): 
+        temporary='./text_file/temporary.txt'
+        if self.localdic != {}:
+            for model in os.listdir('./house_model'):
+                if model.startswith(self.localdic['location']):         #確認區域名稱在model資料夾找的到
+                    loaded_model = joblib.load('./house_model/'+self.localdic['location']+'_model')     #載入模型
+                    test = np.array([self.localdic['age'],self.localdic['size'],self.format_number])       #屋齡/坪數/間數
+                    test = test.reshape(1,-1)
+                    poly_reg =PolynomialFeatures(degree=2)
+                    test_ploy =poly_reg.fit_transform(test)
+                    pred = loaded_model.predict(test_ploy)
+                    if int(self.localdic['price']) < int(pred)*4:      #價格低於預測標準(目前pred*4)，放入第一優先index
+                        index.append(self.localdic)
+                        index1.append(self.localdic)
+                        self.new_mg.save('../house_web/static/screenshot/'+self.filename+'1-'+str(num1+len(index1))+'.jpg')   #儲存圖片連結
+                    elif int(self.localdic['price']) < int(pred)*6:    #價格低於預測標準(目前pred*6)，放入第二優先index
+                        index.append(self.localdic)
+                        index2.append(self.localdic)
+                        self.new_mg.save('../house_web/static/screenshot/'+self.filename+'2-'+str(num2+len(index2))+'.jpg')   #儲存圖片連結
+                    else:
+                        self.localdic = {}
+                    break
+                else:
+                    continue
+        os.remove(self.filename+".jpg")
         os.remove(temporary)
    
     def makefile(self,index,index1,index2):        #將txt轉成另存成json
         goal = self.path+self.filename+".txt" 
-        jsonfile1 = './house_web/static/data/'+self.filename+'1.json'
-        jsonfile2 = './house_web/static/data/'+self.filename+'2.json'
-        with open(goal,"a",encoding='utf-8') as g:
+        jsonfile1 = '../house_web/static/data/'+self.filename+'1.json'
+        jsonfile2 = '../house_web/static/data/'+self.filename+'2.json'
+        with open(goal,"a",encoding='utf-8') as g:      #全部資訊txt檔
             json.dump(index,g,indent=4,ensure_ascii=False)
-        with open(jsonfile1,"a",encoding='utf-8') as j:
-            json.dump(index1,j,indent=4,ensure_ascii=False)
-        with open(jsonfile2,"a",encoding='utf-8') as j:
-            json.dump(index2,j,indent=4,ensure_ascii=False)
+        if os.path.isfile(jsonfile1):               #第一優先json檔
+            with open(jsonfile1,'r',encoding='utf-8') as j:
+                jslist1 = json.load(j)
+                for i in index1:
+                    jslist1.append(i)
+            with open(jsonfile1,"w",encoding='utf-8') as j:
+                json.dump(jslist1,j,indent=4,ensure_ascii=False)
+        else:
+            with open(jsonfile1,"w",encoding='utf-8') as j:
+                json.dump(index1,j,indent=4,ensure_ascii=False)
+        if os.path.isfile(jsonfile2):               #第二優先json檔
+            with open(jsonfile2,'r',encoding='utf-8') as j:
+                jslist2 = json.load(j)
+                for i in index2:
+                    jslist2.append(i)
+            with open(jsonfile2,"w",encoding='utf-8') as j:
+                json.dump(jslist2,j,indent=4,ensure_ascii=False)
+        else:
+            with open(jsonfile2,"w",encoding='utf-8') as j:
+                json.dump(index2,j,indent=4,ensure_ascii=False)
